requirements.txt
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.2
distlib==0.3.0
filelock==3.0.12
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
pytz==2019.3
six==1.14.0
urllib3==1.25.8
virtualenv==20.0.8
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
youtube-dl==2020.3.8